requirements.txt
numpy
pandas
Pillow
plotly
scipy
streamlit
